Move ``process_tokens`` to ``_MessageStateHandler``
